+++ frontend/InstancesButtons.tsx
@@ -2,10 +2,10 @@ import type { Component } from "solid-js";
 
 import { Button, notificationService, Flex, Tooltip } from "@hope-ui/solid";
 
-import { refetchInstances } from "./store";
+import { clearMessages, refetchInstances } from "./store";
 
 const CLEAR_INFO = `\
-Clear both webxdc-dev server state as well as client state.
+Reset both webxdc-dev server state as well as client state.
 This wipes out any localStorage and sessionStorage on each client, and reloads them.`;
 
 const InstancesButtons: Component<{
@@ -26,8 +26,9 @@ const InstancesButtons: Component<{
 
   const handleClear = async () => {
     await fetch(`/clear`, { method: "POST" });
+    clearMessages();
     notificationService.show({
-      title: `Clearing state of dev server & instances`,
+      title: `Resetting state of dev server & instances`,
     });
   };
 
@@ -35,7 +36,7 @@ const InstancesButtons: Component<{
     <Flex direction="row" justifyContent="flex-start" gap="$3">
       <Button onClick={handleAddInstance}>Add Instance</Button>
       <Tooltip label={CLEAR_INFO}>
-        <Button onClick={handleClear}>Clear</Button>
+        <Button onClick={handleClear}>Reset</Button>
       </Tooltip>
     </Flex>
   );